🔨Refactor: #12 vite 프로젝트 구조 재설계
0. 기존 아토믹 패턴을 개량함
1. 도메인형과 계층형을 동시에 사용하도록 조정
2. 조직 이하 컴포넌트는 계층형으로 사용하여 재사용성 확장
3. 템플릿 이상 컴포넌트는 재사용될 가능성이 낮기 때문에 도메인형으로 구축함
4. 기존에 분산된 공용 디렉터리들을 common으로 하나로 묶음
5. 동적 할당은 json으로 처리하고 정적 할당은 enum으로 처리하도록 변경

추가: 최종적으로 프로젝트 구조 초안이 나왔고 이를 통해 인증 페이지 작업 시작, 추후 RTK를 도입하면서 추가적으로 개편 에정
