+++ .gitignore
@@ -121,9 +121,6 @@ logs
 **/*.backup.*
 **/*.back.*
 
-node_modules
-bower_components
-
 *.sublime*
 
 psd
