+++ app/src/App.tsx
@@ -1,7 +1,11 @@
-import AppRouter from "./(router)/_app.router";
+import AppRouter from "./router/_app.router";
 
 function App(): JSX.Element {
   return <AppRouter />;
 }
 
 export default App;
+
+//필요시 types, interface, utils, hooks, api, styles, components, storage, config, features, repositories, logic 등을 정의
+//Type, generic, enum, const enum
+// RTK, query, thunk
