+++ app/src/common/constant/image.ts
@@ -1,0 +1,9 @@
+import kokLogoWhite from "../assets/images/kok_white.svg";
+import kokLogoBlack from "../assets/images/kok_black.svg";
+
+const images = {
+  kokLogoWhite: kokLogoWhite,
+  kokLogoBlack: kokLogoBlack,
+};
+
+export default images;
