--- app/src/common/constants/image.ts
@@ -1,9 +1,0 @@
-import kokLogoWhite from "../../assets/images/kok_white.svg";
-import kokLogoBlack from "../../assets/images/kok_black.svg";
-
-const images = {
-  kokLogoWhite: kokLogoWhite,
-  kokLogoBlack: kokLogoBlack,
-};
-
-export default images;
