--- app/src/components/molecules/molecules.index.ts
@@ -1,18 +1,0 @@
-import SignImage from "./sign.image/auth.image";
-import SignField from "./sign.field/auth.field";
-import SignButton from "./sign.button/sign.button";
-import VerificationTitle from "./verification.title/verification.title";
-import VerificationContent from "./verification.content/verification.content";
-import TeamText from "./team.text/team.text";
-import SignText from "./sign.text/sign.text";
-
-const M = {
-  SignImage,
-  SignField,
-  SignButton,
-  SignText,
-  VerificationTitle,
-  VerificationContent,
-  TeamText,
-};
-export default M;
