--- app/src/components/molecules/sign.button/sign.button.tsx
@@ -1,6 +1,0 @@
-import S from "./style";
-
-function SignButton(): JSX.Element {
-  return <S.SignButton>시작하기</S.SignButton>;
-}
-export default SignButton;
