--- app/src/components/molecules/sign.button/style.ts
@@ -1,6 +1,0 @@
-import { Button, styled } from "@mui/material";
-
-const SignButton = styled(Button)({});
-
-const S = { SignButton };
-export default S;
