--- app/src/components/molecules/sign.field/auth.field.tsx
@@ -1,6 +1,0 @@
-import S from "./style";
-
-function SignField(): JSX.Element {
-  return <S.SignField helperText="회원가입" />;
-}
-export default SignField;
