--- app/src/components/molecules/sign.field/style.ts
@@ -1,8 +1,0 @@
-import { styled, TextField } from "@mui/material";
-
-const SignField = styled(TextField)({});
-
-const S = {
-  SignField,
-};
-export default S;
