--- app/src/components/molecules/sign.image/auth.image.tsx
@@ -1,7 +1,0 @@
-import images from "../../../common/constants/image";
-import S from "./style";
-
-function SignImage(): JSX.Element {
-  return <S.SignImage src={images.kokLogoWhite} alt="test" />;
-}
-export default SignImage;
