--- app/src/components/molecules/sign.image/style.ts
@@ -1,6 +1,0 @@
-import { styled } from "@mui/material";
-
-const SignImage = styled("img")({});
-
-const S = { SignImage };
-export default S;
