--- app/src/components/molecules/sign.text/sign.text.tsx
@@ -1,6 +1,0 @@
-import S from "./style";
-
-function SignText(): JSX.Element {
-  return <S.SignText>계정을 잊으셨나요?</S.SignText>;
-}
-export default SignText;
