--- app/src/components/molecules/sign.text/style.ts
@@ -1,6 +1,0 @@
-import { Box, styled } from "@mui/material";
-
-const SignText = styled(Box)({});
-
-const S = { SignText };
-export default S;
