--- app/src/components/molecules/team.text/style.ts
@@ -1,6 +1,0 @@
-import { Box, styled } from "@mui/material";
-
-const TeamText = styled(Box)({});
-
-const S = { TeamText };
-export default S;
