--- app/src/components/molecules/team.text/team.text.tsx
@@ -1,6 +1,0 @@
-import S from "./style";
-
-function TeamText(): JSX.Element {
-  return <S.TeamText>Team. IGoose</S.TeamText>;
-}
-export default TeamText;
