--- app/src/components/molecules/verification.content/style.ts
@@ -1,6 +1,0 @@
-import { Box, styled } from "@mui/material";
-
-const VerificationContent = styled(Box)({});
-
-const S = { VerificationContent };
-export default S;
