--- app/src/components/molecules/verification.content/verification.content.tsx
@@ -1,6 +1,0 @@
-import S from "./style";
-
-function VerificationContent(): JSX.Element {
-  return <S.VerificationContent>Test</S.VerificationContent>;
-}
-export default VerificationContent;
