--- app/src/components/molecules/verification.title/style.ts
@@ -1,6 +1,0 @@
-import { Box, styled } from "@mui/material";
-
-const VerificationTitle = styled(Box)({});
-
-const S = { VerificationTitle };
-export default S;
