--- app/src/components/molecules/verification.title/verification.title.tsx
@@ -1,6 +1,0 @@
-import S from "./style";
-
-function VerificationTitle(): JSX.Element {
-  return <S.VerificationTitle>Test</S.VerificationTitle>;
-}
-export default VerificationTitle;
