--- app/src/components/organisms/organism.index.ts
@@ -1,4 +1,0 @@
-import SignBody from "./sign.body/sign";
-
-const O = { SignBody };
-export default O;
