--- app/src/components/organisms/sign.body/style.ts
@@ -1,6 +1,0 @@
-import { Box, styled } from "@mui/material";
-
-const SignBody = styled(Box)({});
-
-const S = { SignBody };
-export default S;
