--- app/src/components/templates/sign/style.ts
@@ -1,6 +1,0 @@
-import { Box, styled } from "@mui/material";
-
-const Container = styled(Box)({});
-
-const S = { Container };
-export default S;
