--- app/src/pages/sign/sign.page.tsx
@@ -1,6 +1,0 @@
-import Sign from "../../components/templates/sign/sign";
-
-function SignPage(): JSX.Element {
-  return <Sign />;
-}
-export default SignPage;
