+++ app/src/router/authRoutes.tsx
@@ -1,0 +1,9 @@
+import { RouteObject } from "react-router-dom";
+
+const authRoutes: RouteObject[] = [
+  {
+    path: "/",
+    element: <>test</>,
+  },
+];
+export default authRoutes;
